add: 店铺每月固定费

+++ shares/models.py
@@ -881,9 +881,7 @@ class ShopFixedFees(models.Model):
     shop_name = models.CharField(max_length=50, blank=True, null=True)
     effect_month = models.DateField(blank=True, null=True)
     fee_name = models.CharField(max_length=50, blank=True, null=True)
-    fee_amount = models.DecimalField(
-        max_digits=18, decimal_places=2, blank=True, null=True
-    )
+    fee_amount = models.IntegerField(blank=True, null=True)
     updated_at = models.DateTimeField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_by = models.ForeignKey(
